Show page works.

+++ db/migrate/20141016143649_add_weighted_score_to_users.rb
@@ -1,5 +1,5 @@
 class AddWeightedScoreToUsers < ActiveRecord::Migration
   def change
-    add_column :users, :weight, :integer, default: 5
+    add_column :users, :weight, :float, default: 5
   end
 end
